Start of SceneComponent.
SceneComponent represents an Entity within a scene (game). It has a
transform, snapshot and live states, and finally a render state. It also
has methods to apply physics.

This is a drastic change to how Spout handles Transforms on Entities. For
now on, all entities are gurranteed to have physics available for use.
Signed-off-by: Chris Sanders <zidane@outlook.com>