Fix core meshing.
The problem stemmed from the fact that the queue system is not the dirty
system. This was done for performance and threading reasons: If the queue
drags out a chunk already wiped of dirty status, we don't want to redo
work. For SERVER mode this works fine, CLIENT mode needs us to manually
set state.

I also addressed some other quirks as well as oversights in grabbing
chunks. PRESNAPSHOT should never grab a chunk with a loadopt of anything
but NO_LOAD.

The only issue remaining now is in SpoutChunkSnapshotGroup, NPEs occur but
the chunks are definately in the queue now.
Signed-off-by: Chris Sanders <zidane@outlook.com>